Refactor PR57 calibration cleanup
Split calibration model fitting into focused modules for types, feature builders, ridge regression, diagnostics, metrics, IDW fallback, and polynomial ridge fitting while preserving the public calibration-models facade.

Extract webcam gaze camera, MediaPipe loading, landmark parsing, and shared types behind smaller modules without changing the hook return shape.

Centralize reading-zone bounds for live display, replay, and export; keep calibration diagnostics behind calibrationLogger; ignore web/tmp artifacts.

Fix cleanup issues from review: enforce the outlier retained-sample floor, remove runtime console assertions, route ridge lambda logs through the logger, remove unused calibration wrapper code, and document remaining follow-up work in the temporary tracker.
